Fixed video archiving
First I fixed the same issue with the paths for the archived videos.
I added a path into the config.py and removed the hardcoding.
I moved it into the /app/data folder aswell, so all data is tidy.
The folderstructer check aswell at runtime.
The docker-compose got some comments for dockerbeginner and were
updated to save the the collection and data persisistant in
./local_persistant (added to gitignore).